Refactor to schema service
